+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and adheres to [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
+- Optimized `ReadReportBasesByLocation` EF query in `ReportQueries` to fetch
+  network users directly by location foreign key
 - Fix, `ApiV1Href` property on `OzdsComponentBase` is now just a string that
   points from root of url to the swagger API
 - update mailkit